ntfsck: A parts of src/utils.c moved to libntfs-3g
To use utils_is_metadata() in libntfs-3g, some functino moved to
libntfs-3g/
