+++ src/utils.h
@@ -30,6 +30,7 @@
 #include "types.h"
 #include "layout.h"
 #include "volume.h"
+#include "lib_utils.h"
 
 #ifdef HAVE_ERRNO_H
 #include <errno.h>
@@ -49,11 +50,7 @@ int utils_attr_get_name(ntfs_volume *vol, ATTR_RECORD *attr, char *buffer, int b
 int utils_cluster_in_use(ntfs_volume *vol, long long lcn);
 int utils_mftrec_in_use(ntfs_volume *vol, MFT_REF mref);
 int check_mftrec_in_use(ntfs_volume *vol, MFT_REF mref, int force);
-int utils_is_metadata(ntfs_inode *inode);
 void utils_dump_mem(void *buf, int start, int length, int flags);
-
-ATTR_RECORD * find_attribute(const ATTR_TYPES type, ntfs_attr_search_ctx *ctx);
-ATTR_RECORD * find_first_attribute(const ATTR_TYPES type, MFT_RECORD *mft);
 
 int utils_valid_device(const char *name, int force);
 ntfs_volume * utils_mount_volume(const char *device, unsigned long flags);
